requirements.txt
flask
requests
transformers
torch
matplotlib
gunicorn
mysql-connector-python